edit currency styles
