--- src/components/Currency/CurrencyTab/index.js
@@ -1,1 +1,0 @@
-export { default } from './CurrencyTab';
